feat: replace all integer ids with long (i know uuid is better) and add readme

See merge request Dkorolkov/mmtback!32

+++ auth-service/src/main/java/org/example/auth/config/KafkaProducerConfig.java
@@ -1,7 +1,7 @@
 package org.example.auth.config;
 
 import org.apache.kafka.clients.producer.ProducerConfig;
-import org.apache.kafka.common.serialization.IntegerSerializer;
+import org.apache.kafka.common.serialization.LongSerializer;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -19,16 +19,16 @@ public class KafkaProducerConfig {
     private String bootstrapAddress;
 
     @Bean
-    public ProducerFactory<Integer, Object> createPersonProducer() {
+    public ProducerFactory<Long, Object> createPersonProducer() {
         Map<String, Object> properties = new HashMap<>();
         properties.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, bootstrapAddress);
-        properties.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, IntegerSerializer.class);
+        properties.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, LongSerializer.class);
         properties.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, JsonSerializer.class);
         return new DefaultKafkaProducerFactory<>(properties);
     }
 
     @Bean
-    public KafkaTemplate<Integer, Object> createPersonKafkaTemplate() {
+    public KafkaTemplate<Long, Object> createPersonKafkaTemplate() {
         return new KafkaTemplate<>(createPersonProducer());
     }
 }
