+++ auth-service/src/main/java/org/example/auth/dto/mmtr/MmtrTokenInfoDto.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class MmtrTokenInfoDto {
     private Long expiration;
     private Boolean isDeleted;
-    private Integer employeeId;
+    private Long employeeId;
     private String tokenInfo;
-    private Integer tokenId;
+    private Long tokenId;
 }
