+++ auth-service/src/main/java/org/example/auth/repository/RefreshTokenInfoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RefreshTokenInfoRepository extends JpaRepository<RefreshTokenInfo, Integer> {
+public interface RefreshTokenInfoRepository extends JpaRepository<RefreshTokenInfo, Long> {
     Optional<RefreshTokenInfo> findRefreshTokenInfoByRefreshToken(String refreshToken);
 
     void deleteRefreshTokenInfoByRefreshToken(String refreshToken);
