+++ course-registration-service/src/main/java/org/example/courseregistration/exception/classes/request/RequestNotFoundException.java
@@ -1,7 +1,7 @@
 package org.example.courseregistration.exception.classes.request;
 
 public class RequestNotFoundException extends RuntimeException {
-    public RequestNotFoundException(Integer requestId) {
+    public RequestNotFoundException(Long requestId) {
         super("Запрос " + requestId + " не найден");
     }
