+++ course-registration-service/src/main/java/org/example/courseregistration/exception/classes/request/RequestNotRejectedException.java
@@ -1,7 +1,7 @@
 package org.example.courseregistration.exception.classes.request;
 
 public class RequestNotRejectedException extends RuntimeException {
-    public RequestNotRejectedException(Integer requestId) {
+    public RequestNotRejectedException(Long requestId) {
         super("Запрос " + requestId + " не был закрыт с отказом");
     }
 }
