+++ course-registration-service/src/main/java/org/example/courseregistration/repository/CourseRegistrationRequestKeeperRepository.java
@@ -8,12 +8,12 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface CourseRegistrationRequestKeeperRepository extends JpaRepository<CourseRegistrationRequestKeeper, Integer> {
-    List<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeepersByRequestId(Integer requestId);
+public interface CourseRegistrationRequestKeeperRepository extends JpaRepository<CourseRegistrationRequestKeeper, Long> {
+    List<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeepersByRequestId(Long requestId);
 
     @Query("SELECT crrk FROM CourseRegistrationRequestKeeper crrk\n" +
             "WHERE crrk.keeperId IN :keeperIds AND crrk.status.status = 'PROCESSING'")
-    List<CourseRegistrationRequestKeeper> findProcessingCourseRegistrationRequestKeepersByKeeperIdIn(@Param("keeperIds") List<Integer> keeperIds);
+    List<CourseRegistrationRequestKeeper> findProcessingCourseRegistrationRequestKeepersByKeeperIdIn(@Param("keeperIds") List<Long> keeperIds);
 
-    Optional<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeeperByRequestIdAndKeeperId(Integer requestId, Integer keeperId);
+    Optional<CourseRegistrationRequestKeeper> findCourseRegistrationRequestKeeperByRequestIdAndKeeperId(Long requestId, Long keeperId);
 }
