+++ course-registration-service/src/main/java/org/example/courseregistration/repository/CourseRepository.java
@@ -1,5 +1,5 @@
 package org.example.courseregistration.repository;
 
 public interface CourseRepository {
-    Boolean existsById(Integer courseId);
+    Boolean existsById(Long courseId);
 }
