+++ course-registration-service/src/main/java/org/example/courseregistration/repository/RejectionReasonRepository.java
@@ -3,5 +3,5 @@ package org.example.courseregistration.repository;
 import org.example.courseregistration.model.RejectionReason;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RejectionReasonRepository extends JpaRepository<RejectionReason, Integer> {
+public interface RejectionReasonRepository extends JpaRepository<RejectionReason, Long> {
 }
