+++ course-service/src/main/java/org/example/course/config/security/RoleService.java
@@ -27,7 +27,7 @@ public class RoleService {
         return false;
     }
 
-    public boolean hasAnyCourseRole(Integer courseId, CourseRoleType role) {
+    public boolean hasAnyCourseRole(Long courseId, CourseRoleType role) {
         PeopleService.Person person = (PeopleService.Person) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         if (role.equals(CourseRoleType.EXPLORER))
             return explorerRepository.findExplorerByPersonIdAndGroup_CourseId(person.getPersonId(), courseId).isPresent();
@@ -35,7 +35,7 @@ public class RoleService {
             return keeperRepository.findKeeperByPersonIdAndCourseId(person.getPersonId(), courseId).isPresent();
     }
 
-    public boolean hasAnyCourseRoleByThemeId(Integer themeId, CourseRoleType role) {
+    public boolean hasAnyCourseRoleByThemeId(Long themeId, CourseRoleType role) {
         return hasAnyCourseRole(
                 courseThemeRepository.findById(themeId)
                         .orElseThrow(() -> new CourseThemeNotFoundException(themeId))
