+++ course-service/src/main/java/org/example/course/dto/keeper/KeeperBaseInfoDto.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class KeeperBaseInfoDto {
-    private Integer personId;
+    private Long personId;
     private String firstName;
     private String lastName;
     private String patronymic;
-    private Integer keeperId;
+    private Long keeperId;
 }
